Fifth commit to edit headings and descriptions for each script
Adding informative descriptions for each script including a header, a short description of what the script does, the author's name and date.
This commit also includes a fix in the preprocessing of the collected data script - shortened the binding of subjects' csvs by calling the files in the rbind functions instead of first creating a temporary df for each file before the binding.